// src/chat/enum/chatResponseType.enum.ts
/**
 * Stable response identifiers shared with the frontend ResponseType enum.
 */
export enum ChatResponseType {
  NEED_COMPANY = 'NeedCompany',
  NEW_GAME = 'NewGame',
  COMING_TO_PLAY = 'ComingToPlay',
  WANT_TO_PLAY = 'WantToPlay',
  YES = 'Yes',
  NO = 'No',
  ONLINE = 'Online',
  LEAVING = 'Leaving',
  BUSY = 'Busy',
  WINNING_STREAK = 'WinningStreak',
  LOSING_STREAK = 'LosingStreak',
  DISCORD = 'Discord',
  LONELY = 'Lonely',
  GOOD_GAME = 'GoodGame',
  BAD_GAME = 'BadGame',
  SKILL_ISSUE = 'SkillIssue',
}